Implemented File System, refactor, ecc
Added a File System and individual file structure and data structures to support them (hash table and dl list)
Improved signal handling (won't deadlock if signal is sent while a request is being completed)
Currently no errors or memory leaks.

+++ API.h
@@ -4,7 +4,7 @@
 typedef struct{
     int func;
     int path_size;
-    double file_size;
+    size_t file_size;
 }SockMsg;
 
 int openConnection(const char* sock_name, int msec, const struct timespec abstime);
